move bindValue & bindValues to CommandPDO class
